// src/modules/textile/reports/Dyeing/grey-fabric-process-challan-report/grey-fabric-process-challan-report-type.ts
export type GreyFabricProcessChallanReportType = {
  COMPANY_NAME?: string;
  COMPANY_ADDRESS?: string;
  DYEING_COMPANY?: string;
  SUPPLIER?: string;
  SUPPLIER_ADDRESS?: string;
  DRIVER_NAME?: string;
  DRIVER_MOBILE?: string;
  ISSUE_CHALLAN?: string;
  ISSUE_DATE?: string;
  SUPERVISOR_NAME?: string;
  SUPERVISOR_MOBILE?: string;
  BUYER_NAME?: string;
  BUYER_NICK_NAME?: string;
  STYLENO?: string;
  JOB_NUMBER?: string;
  PONO?: string;
  FABRIC?: string;
  YARN_LOT?: string;
  YARN_BRAND?: string;
  YARN_NAME?: string;
  COLORNAME?: string;
  TREATMENT?: string;
  FIN_DIA?: string;
  GSM?: string;
  STITCH_LENGTH?: string;
  ISSUE_QTY_ROLL?: number;
  ISSUE_QTY_KG?: number;
  REMARKS?: string;
  PREPARED_BY?: string;
  UNIQUE_ROW?: string;
  GREY_SHAPE?: string;
  FIN_REQ_SHAPE?: string;
  SYMBOLIC_NAME?: string;
  DTL_REMARKS?: string;
  VEHICLE_NO?: string;
};